Updated transaction and delete functionality

+++ XpensAuditor/app/src/main/java/com/xa/xpensauditor/SumTrans.java
@@ -1,20 +1,23 @@
 package com.xa.xpensauditor;
 
+import android.icu.number.Precision;
+
+import java.text.DecimalFormat;
 import java.util.Collection;
 import java.util.Iterator;
 
 public class SumTrans {
-    public Integer sum;
+    public double sum;
 
     public SumTrans(){
-        sum=0;
+        sum=0.00;
     }
 
-    public Integer computeSum(Collection<String> str){
+    public Double computeSum(Collection<String> str){
         for(Iterator<String> entry = str.iterator(); entry.hasNext();){
 
             String value = entry.next();
-            sum = sum+ Integer.parseInt(value);
+            sum = sum+ Double.parseDouble((value));
             //Log.wtf("afSDFsdfEWSFSDcWEESFsedfSEDFwesedf",sum.toString());
 
         }
